refactor(control): split nfqueue transitions
